message

+++ src/bookings/booking.service.ts
@@ -10,11 +10,22 @@ import { HttpStatus } from "@nestjs/common";
 export class BookingService {
   constructor(private prisma: PrismaService) {}
   async createBooking(createBookingDto: CreateBookingDto, id: string, req) {
+    const router = await this.prisma.router.findUnique({
+      where: {
+        domain: id,
+      },
+    });
+    if (!router) {
+      throw new HttpException(
+        "this router doesn't exist",
+        HttpStatus.BAD_REQUEST,
+      );
+    }
     try {
       const booking = await this.prisma.booking.create({
         data: { ...createBookingDto, bookingID: id, userId: req.user.userId },
       });
-      return booking;
+      return { ...booking, message: "booking created successfully" };
     } catch (err) {
       return err;
     }
@@ -22,7 +33,7 @@ export class BookingService {
   async updateBooking(
     updateBookingDto: CreateBookingDto,
     id: string,
-    router,
+    router: string,
     req,
   ) {
     try {
@@ -37,6 +48,17 @@ export class BookingService {
           HttpStatus.BAD_REQUEST,
         );
       }
+      const routerExist = await this.prisma.router.findUnique({
+        where: {
+          domain: router,
+        },
+      });
+      if (!routerExist) {
+        throw new HttpException(
+          "this router doesn't exist",
+          HttpStatus.BAD_REQUEST,
+        );
+      }
       const booking = await this.prisma.booking.update({
         where: {
           id: parseInt(id),
@@ -47,43 +69,66 @@ export class BookingService {
           bookingID: router,
         },
       });
-      return booking;
+      return { ...booking, message: "booking updated successfully" };
     } catch (err) {
       return err;
     }
   }
-  async allBookings(routerId) {
+  async allBookings(routerId: string) {
     try {
       const bookings = await this.prisma.booking.findMany({
         where: {
           bookingID: routerId,
         },
       });
-      return bookings;
+      if (bookings.length === 0) {
+        throw new HttpException(
+          "booking doesn't exist",
+          HttpStatus.BAD_REQUEST,
+        );
+      }
+      return { ...bookings, message: "bookings fetched successfully" };
     } catch (err) {
       return err;
     }
   }
-  async bookingById(id) {
+  async bookingById(id: string) {
     try {
       const booking = await this.prisma.booking.findUnique({
         where: {
           id: parseInt(id),
         },
       });
-      return booking;
+      if (!booking) {
+        throw new HttpException(
+          "this booking doesn't exist",
+          HttpStatus.BAD_REQUEST,
+        );
+      }
+      return { ...booking, message: "booking fetched successfully" };
     } catch (err) {
       return err;
     }
   }
   async deleteBooking(id: string) {
     try {
+      const booking = await this.prisma.booking.findUnique({
+        where: {
+          id: parseInt(id),
+        },
+      });
+      if (!booking) {
+        throw new HttpException(
+          "this booking doesn't exist",
+          HttpStatus.BAD_REQUEST,
+        );
+      }
       await this.prisma.booking.delete({
         where: {
           id: parseInt(id),
         },
       });
-      return "booking hase been canceled";
+      return { message: "booking canceled" };
     } catch (err) {
       return err;
     }
